ScePT/poses/Pose_Estimation_main/requirements.txt
absl-py==0.15.0
addict==2.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
clang==5.0
click==8.1.3
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flake8==4.0.1
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.1
h5py==3.1.0
idna==3.3
imageio==2.19.2
immutabledict==2.2.1
importlib-metadata==4.11.3
importlib-resources==5.7.1
install==1.3.5
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-packaging==0.12.2
jupyter-server==1.18.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.1
kaleido==0.2.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.4.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.19.5
oauthlib==3.2.0
open3d==0.15.2
opencv-python==4.5.5.64
OpenEXR==1.3.8
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
setuptools
plotly==5.8.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
pyntcloud==0.3.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
tensorflow-graphics==2021.12.3
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.11.0
torch==1.12.1
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torch-spline-conv==1.2.1
torchgeometry==0.1.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
trimesh==3.12.0
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.9
Wand==0.6.7
wandb==0.12.17
waymo-open-dataset-tf-2-6-0 
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wrapt==1.12.1
zipp==3.8.0
